fix(Nirjas): fix directory name and import statements
