+++ nirjas/languages/perl.py
@@ -20,18 +20,18 @@ License along with this library; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 '''
 
-from extractor.binder import *
+from nirjas.binder import *
 
-def javaExtractor(file):
+def perlExtractor(file):
     result = CommentSyntax()
-    result1 = result.doubleSlash(file)
-    result2 = result.slashStar(file)
+    result1 = result.hash(file)
+    result2 = result.beginCut(file)
     result4 = contSingleLines(result1)
     file = file.split("/")
     output = {
         "metadata": [{
         "filename": file[-1],
-        "lang": "Java",
+        "lang": "Perl",
         "total_lines": result1[1],
         "total_lines_of_comments": result1[3]+result2[3],
         "blank_lines": result1[2],
@@ -56,33 +56,33 @@ def javaExtractor(file):
     if result2:
         for idx,i in enumerate(result2[0]):
             output['multi_line_comment'].append({"start_line": result2[0][idx], "end_line": result2[1][idx], "comment": result2[2][idx]})
-
+        
 
     return output
 
 
-def javaSource(file, newFile: str):
+def perlSource(file, newFile: str):
     closingCount = 0
     copy = True
     with open(newFile, 'w+') as f1:
         with open(file) as f:
             for lineNumber, line in enumerate(f, start=1):
-                if line.strip() == '/*':
+                if line.strip() == '=begin':
                     closingCount+=1
                     copy = False
                     if closingCount%2 == 0:
                         copy = True
 
-                if line.strip() == '*/':
+                if line.strip() == '=cut':
                     closingCount+=1
                     copy = False
                     if closingCount%2 == 0:
                         copy = True
 
                 if copy:
-                    if line.strip() != '/*' and line.strip() != '*/':
+                    if line.strip() != '=begin' and line.strip() != '=cut':
                         Templine = line.replace(" ","")
-                        if Templine[0:2] != "//":            # Syntax for single line comment
+                        if Templine[0] != "#":            # Syntax for single line comment
                             f1.write(line)
     f.close()
     f1.close()
