+++ nirjas/languages/text.py
